feat: moving handlers into events folder
