+++ src/handler/events/dispatchers.ts
@@ -6,19 +6,19 @@ import type {
     ModalSubmitCommand,
     SelectMenuCommand,
     SlashCommand,
-} from '../../structures/module';
-import Context from '../../structures/context';
-import type { SlashOptions } from '../../../types/handler';
-import { asyncResolveArray } from '../../utilities/asyncResolveArray';
-import { controller } from '../../sern';
+} from '../structures/module';
+import Context from '../structures/context';
+import type { SlashOptions } from '../../types/handler';
+import { asyncResolveArray } from '../utilities/asyncResolveArray';
+import { controller } from '../sern';
 import type { ButtonInteraction, ModalSubmitInteraction } from 'discord.js';
 import type { SelectMenuInteraction } from 'discord.js';
 import type { AutocompleteInteraction, ChatInputCommandInteraction } from 'discord.js';
-import { isAutocomplete } from '../../utilities/predicates';
+import { isAutocomplete } from '../utilities/predicates';
 import type { Interaction } from 'discord.js';
 import type { UserContextMenuCommandInteraction } from 'discord.js';
 import type { MessageContextMenuCommandInteraction } from 'discord.js';
-import { SernError } from '../../structures/errors';
+import { SernError } from '../structures/errors';
 
 export function applicationCommandDispatcher(interaction: Interaction) {
     if (isAutocomplete(interaction)) {
